# docs/swagger.yaml
basePath: /api/v1
definitions:
  dto.AddressRequest:
    properties:
      city:
        example: Moscow
        type: string
      country:
        example: Russia
        type: string
      street:
        example: Vernadskogo
        type: string
    required:
    - city
    - country
    - street
    type: object
  dto.AddressResponse:
    properties:
      address_id:
        example: b1c4b160-b9e9-4f0c-83ec-9119e1f467f9
        type: string
      city:
        example: Moscow
        type: string
      country:
        example: Russia
        type: string
      street:
        example: Vernadskogo
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      error:
        example: internal server error
        type: string
    type: object
...
  dto.SuccessResponse:
    properties:
      status:
        example: created
        type: string
    type: object
  dto.SupplierListResponse:
    properties:
      data:
        items:
          $ref: '#/definitions/dto.SupplierResponse'
        type: array
      limit:
        example: 20
        type: integer
      offset:
        example: 0
        type: integer
      total:
        example: 5
        type: integer
    type: object
  dto.SupplierRequest:
    properties:
      address:
        $ref: '#/definitions/dto.AddressRequest'
      name:
        example: Super Keyboard Supplier
        type: string
      phone_number:
        example: "+79995351675"
        type: string
    required:
    - address
    - name
    - phone_number
    type: object
  dto.SupplierResponse:
    properties:
      address:
        $ref: '#/definitions/dto.AddressResponse'
      name:
        example: Super Keyboard Supplier
        type: string
      phone_number:
        example: "+79995351675"
        type: string
      supplier_id:
        example: b1c4b160-b9e9-4f0c-83ec-9119e1f467f9
        type: string
    type: object
info:
  contact: {}
  description: REST API for shop service
  title: Shop API
  version: "1.0"
paths:
...
  /suppliers:
    get:
      description: Get suppliers list with optional params
      parameters:
      - default: 100
        description: Limit
        in: query
        maximum: 100
        minimum: 1
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        minimum: 0
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SupplierListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get suppliers with params
      tags:
      - suppliers
    post:
      consumes:
      - application/json
      description: Create new supplier
      parameters:
      - description: Supplier data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SupplierRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Create supplier
      tags:
      - suppliers
  /suppliers/{id}:
    delete:
      description: Delete supplier by id
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Delete supplier
      tags:
      - suppliers
    get:
      description: Get supplier by id
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SupplierResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Get supplier
      tags:
      - suppliers
  /suppliers/{id}/address:
    patch:
      consumes:
      - application/json
      description: Change address of supplier
      parameters:
      - description: Supplier ID
        in: path
        name: id
        required: true
        type: string
      - description: Address data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.AddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Change supplier address
      tags:
      - suppliers
swagger: "2.0"